IANA Service Name and Transport Protocol Port Number Registry

This is the dataset of IANA service names and their assigned ports and transport protocols - along with related metadata.

service_name: The service name for the port assignment

port_number: The ports assigned to that service. This can be individual ports, or a range.

transport_protocol: The transport protocol(s) of the port assignment - [dccp|sctp|tcp|udp]

description: An explanation of the port assignment

assignee: the name of the individual or organisation to whom the assignment is made

contact: the name of the individual or organisation who serves as the contact person for the assignment.

registration_date. The date the assignment was registered on. This may be empty, in the case of early assignments; otherwise, it is stored in the form "YYYY-MM".

modification_date: The date of any modification to the assignment. Same format as registration_date

reference: A description of (or a reference to a document describing) the protocol or application using this port

known_unauthorised_uses: A list of uses by applications or organizations who are not the assignee

assignment_notes. Indications of owner/name change, or any other assignment process issue

data(iana_ports)

Format

A data frame with 13,659 rows and 12 variables

Note

Last updated 2015-05-02

References